Some profiler refactoring (#2649)
The primary goal of this PR is to make the profiler code structure
clearer top-down (packages + public classes).

The proposed structure of packages + public classes in this PR is:

* ...profiler
  * ProfilingInitializer
  * Profiler
* ...profiler.config
  * ConfigPollingInit
  * ProfilerConfiguration
* ...profiler.service
  * ArtifactAcceptedResponse
  * BlobAccessPass
  * ServiceProfilerClient
* ...profiler.trigger
  * AlertingSubsystemInit
  * AlertTriggerSpanProcessor
* ...profiler.upload
  * ServiceProfilerIndex
  * UploadFinishArgs
  * UploadListener
  * UploadService

This also consolidates the above profiler classes from three modules
into just the `agent-tooling` module, eliminating
`agent-service-profiler` and `agent-profiler-api`.

The PR also does a few other things in the profiler code:
* Introduces `@AutoValue` (and builders) in several places where they
are helpful
* Replaces Moshi json library usage with Jackson json library usage
* Rewrites service tests to use Azure SDK "playback" http responses

+++ agent/agent-tooling/src/main/java/com/microsoft/applicationinsights/agent/internal/configuration/Configuration.java
@@ -1371,7 +1371,7 @@ public class Configuration {
 
   public static class RequestTriggerThrottling {
     public RequestTriggerThrottlingType type = RequestTriggerThrottlingType.FIXED_DURATION_COOLDOWN;
-    public long value = 60; // in seconds
+    public int value = 60; // in seconds
   }
 
   public enum RequestTriggerType {
@@ -1385,7 +1385,7 @@ public class Configuration {
     public RequestAggregation aggregation = new RequestAggregation();
     public RequestTriggerThreshold threshold = new RequestTriggerThreshold();
     public RequestTriggerThrottling throttling = new RequestTriggerThrottling();
-    public long profileDuration = 30; // in s
+    public int profileDuration = 30; // in s
   }
 
   public static class ProfilerConfiguration {
